Attach runtime function definitions to runtimes and handle deletion
