+++ db/migrate/20250112173532_create_runtime_function_definitions.rb
@@ -4,10 +4,12 @@ class CreateRuntimeFunctionDefinitions < Code0::ZeroTrack::Database::Migration[1
   def change
     create_table :runtime_function_definitions do |t|
       t.references :return_type, null: true, foreign_key: { to_table: :data_types, on_delete: :restrict }
-      t.references :namespace, null: false, index: false, foreign_key: { on_delete: :cascade }
+      t.references :runtime, null: false, index: false, foreign_key: { on_delete: :cascade }
       t.text :runtime_name, null: false, limit: 50
 
-      t.index %i[namespace_id runtime_name], unique: true
+      t.datetime_with_timezone :removed_at, null: true
+
+      t.index %i[runtime_id runtime_name], unique: true
 
       t.timestamps_with_timezone
     end
